Change iotools structure for nonregression test folder
